src/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
requests
pydantic-settings
pyyaml
apscheduler
python-multipart
ruff
pytest
httpx
